feat: agregar cobertura de pruebas con JaCoCo al 60%+ en todos los módulos
- Agregar JaCoCo a pom.xml de ms-ventas, ms-datos-org, ms-indicadores y bff
- Agregar 11 nuevas clases de test: controladores, repositorios, servicios y manejadores de excepciones
- ReporteServiceTest (6 tests), VentaRepositoryImplTest (8 tests)
- Controller tests: Venta, Producto, Sucursal, Reporte, Empleado, Departamento, Indicador
- GlobalExceptionHandlerTest en los 3 microservicios
- Mejorar SucursalServiceTest con test de generarDatosMasivos
- ms-ventas: 65.5% | ms-datos-org: 70% | ms-indicadores: 68.6% | bff: 67%

+++ ms-ventas/src/main/java/com/grupocordillera/ms_ventas/service/ReporteService.java
@@ -38,7 +38,7 @@ public class ReporteService {
         return new ResumenVentasDTO(
                 (Long) resumen.get("totalVentas"),
                 (BigDecimal) resumen.get("montoTotal"),
-                (BigDecimal) resumen.get("promedioVenta")
+                BigDecimal.valueOf((Double) resumen.get("promedioVenta"))
         );
     }
